Reworking the auth API server to have the ability to update IP, sync cash types, ban/log user HWID/MAC/IP and ability to set and change PIC. Also added an author tag to HexUtils as it was made by Zygon and i noticed i accidentally removed it there.

+++ CenterServer/src/login/LoginServerSocket.java
@@ -34,8 +34,8 @@ import util.HexUtils;
  */
 public class LoginServerSocket extends Socket {
 
-    public ConcurrentHashMap<String, Integer> mReservedCharacterNames = new ConcurrentHashMap<>();
-    public HashMap<Integer, Account> mAccountStorage = new HashMap<>();
+    public ConcurrentHashMap<String, Long> mReservedCharacterNames = new ConcurrentHashMap<>();
+    public HashMap<Long, Account> mAccountStorage = new HashMap<>();
 
     public LoginServerSocket(Channel channel, int uSeqSend, int uSeqRcv) {
         super(channel, uSeqSend, uSeqRcv);
